fix(tests): run the AgentOutput suites on the headless session thread
Merging #397 turned main red: Unit Tests (ubuntu-latest) failed twice
with System.InvalidOperationException "The calling thread cannot access
this object because a different thread owns it" thrown inside
Avalonia.Headless.XUnit.AvaloniaTestRunner.Run, in the tab strip's
tests - a class #397 never touched. Reverting #397 on a branch made
ubuntu green again, so the trigger was ours.

Cause: MarkdownRendererTests and FenceBodyTests are plain [Fact] classes
that build real Avalonia visuals. Avalonia's ambient state - the media
context above all - has thread affinity to the headless session's
dispatch thread, so constructing visuals from a plain [Fact] touches
state the thread does not own and corrupts it for every [AvaloniaFact]
suite sharing the process. Those suites then fail non-deterministically,
in unrelated classes, in whatever order the run happened to take. This
was latent before #397; #397 added enough off-session work to tip it.
AvaloniaBootLocatorHygieneTests.TheAmbientMediaContextBelongsToTheSession-
DispatchThread is the guard, and it fails locally on the merged tree.

- Both classes move to [AvaloniaFact]/[AvaloniaTheory]. The comment
  claiming they "run without a headless application" was the invariant
  being violated, so it is replaced with why the session is required.
- The link test synthesized a Pointer and raised PointerPressedEventArgs
  from a plain [Fact], never disposing it. It moves to
  AgentOutputPanelTests, which owns a session, and disposes the pointer
  so a pressed pointer cannot leak capture into the pointer-driven tab
  strip tests.

Verified by the interaction that reproduces it, which a full local run
cannot reach (#81 aborts the host on the real-PTY window tests):
AgentOutput + VerticalTabStrip + TabRunningCommand + the hygiene guard
crashed in DragReorder_Commit_HorizontalStrip with the pointer test
alone moved, and passes 169/169 with both parts applied. The same tab
strip tests pass 33/33 on their own, which is what isolates the
contamination to this suite.
